fix inefficient SU members fetching
(and just grab webgroups as required)
