+++ votes/templates/votes/ticket.html
@@ -19,10 +19,6 @@
 {% block body %}
     <h2>Create Ticket by...</h2>
     <p>
-        <a href="{% url "votes:tickets_date" %}"
-           class="btn btn-outline-success btn-block">
-            Membership cutoff date
-        </a>
         <a href="{% url "votes:tickets_id" %}"
            class="btn btn-outline-success btn-block">
             Manual ID list
